Trade (import/export): cleanup
