+++ src/apps/chat/trade/ImportChats.tsx
@@ -8,11 +8,12 @@ import type { ChatGptSharedChatSchema } from '~/modules/sharing/import.chatgpt';
 import { apiAsync } from '~/modules/trpc/trpc.client';
 
 import { Brand } from '~/common/brand';
+import { InlineError } from '~/common/components/InlineError';
 import { OpenAIIcon } from '~/common/components/icons/OpenAIIcon';
 import { createDConversation, createDMessage, DMessage, useChatStore } from '~/common/state/store-chats';
 
 import { ImportedOutcome, ImportOutcomeModal } from './ImportOutcomeModal';
-import { restoreDConversationsFromJSON } from './trade.json';
+import { loadAllConversationsFromJson } from './trade';
 
 
 export type ImportConfig = { dir: 'import' };
@@ -49,7 +50,7 @@ export function ImportConversations(props: { onClose: () => void }) {
       try {
         const fileString = await blob.text();
         const fileObject = JSON.parse(fileString);
-        restoreDConversationsFromJSON(fileName, fileObject, outcome);
+        loadAllConversationsFromJson(fileName, fileObject, outcome);
       } catch (error: any) {
         outcome.conversations.push({ success: false, fileName, error: `Invalid file: ${error?.message || error?.toString() || 'unknown error'}` });
       }
@@ -145,7 +146,10 @@ export function ImportConversations(props: { onClose: () => void }) {
     {/* [chatgpt] data & controls */}
     {chatGptEdit && <Sheet variant='soft' color='primary' sx={{ display: 'flex', flexDirection: 'column', borderRadius: 'md', p: 1, gap: 1 }}>
 
-      <OpenAIIcon sx={{ mx: 'auto', my: 1 }} />
+      <Box sx={{ display: 'flex', flexDirection: 'row', alignItems: 'center', gap: 1 }}>
+        <InlineError error='Note: OpenAI seems to be blocking importing shared chats at the moment. Proceeding is likely to result in an HTTP error.' severity='danger' />
+        <OpenAIIcon sx={{ mx: 'auto', my: 1 }} />
+      </Box>
 
       <FormControl>
         <FormLabel>
